Refactored Tests to Reduce Duplication
